+++ packages/headless/src/api/analytics/analytics.test.ts
@@ -13,6 +13,7 @@ import {getSearchInitialState} from '../../features/search/search-state';
 import {getPipelineInitialState} from '../../features/pipeline/pipeline-state';
 import {getSearchHubInitialState} from '../../features/search-hub/search-hub-state';
 import {buildMockFacetResponse} from '../../test/mock-facet-response';
+import {buildMockAnalyticsConfiguration} from '../../test/mock-analytics-configuration';
 
 describe('analytics', () => {
   const logger = pino({level: 'silent'});
@@ -142,6 +143,31 @@ describe('analytics', () => {
       };
       const provider = new AnalyticsProvider(state);
       expect(provider.getOriginLevel1()).toEqual(searchHub);
+    });
+
+    describe('#getOriginLevel2', () => {
+      it('when an originLevel2 is not set, #getOriginLevel2 returns "default"', () => {
+        const state: StateNeededByAnalyticsProvider = {
+          configuration: getConfigurationInitialState(),
+        };
+        const provider = new AnalyticsProvider(state);
+
+        expect(provider.getOriginLevel2()).toBe('default');
+      });
+
+      it('when an originLevel2 is set, #getOriginLevel2 returns the value', () => {
+        const originLevel2 = 'youtube';
+        const state: StateNeededByAnalyticsProvider = {
+          configuration: {
+            ...getConfigurationInitialState(),
+            analytics: buildMockAnalyticsConfiguration({originLevel2}),
+          },
+        };
+
+        const provider = new AnalyticsProvider(state);
+
+        expect(provider.getOriginLevel2()).toBe(originLevel2);
+      });
     });
 
     it('when a locale search parameter is configured, #getLanguage returns the correct value', () => {